requirements.txt
black==25.11.0
flake8==7.3.0
Flask==3.1.2
gunicorn==23.0.0
isort==7.0.0
mongomock==4.3.0
pre-commit==4.5.0
pytest==9.0.1
pytest-flask==1.3.0
python-dotenv==1.2.1
